vavae/requirements.txt
# Minimal VA-VAE Requirements - Essential packages only
# ====================================================

# Core ML/DL Frameworks  
pytorch-lightning==2.4.0
transformers==4.48.0
timm==1.0.12
diffusers==0.32.1

# Critical compatibility fix - DO NOT CHANGE
numpy==1.26.4

# Computer Vision
opencv-python==4.9.0.80
pillow>=11.0.0
imageio>=2.30.0
albumentations>=2.0.0

# Mathematical Computing
scipy>=1.15.0
matplotlib>=3.10.0
einops>=0.8.0

# Data & Configuration
pandas>=2.3.0
omegaconf>=2.3.0
PyYAML>=6.0.0
tqdm>=4.65.0

# Training & Evaluation
torchmetrics>=1.7.0
pytorch-fid>=0.3.0
tensorboard>=2.16.0

# Utilities
huggingface-hub>=0.33.0
safetensors>=0.5.0
accelerate>=1.8.0